+++ gplearn/tests/test_genetic.py
@@ -845,6 +845,56 @@ def test_more_verbose_output():
     assert_equal(0, n_lines % 10)
 
 
+def test_parallel_train():
+    """Check predictions are the same for different n_jobs"""
+
+    # Check the regressor
+    ests = [
+        SymbolicRegressor(population_size=100, generations=4, n_jobs=n_jobs,
+                          random_state=0).fit(boston.data[:100, :],
+                                              boston.target[:100])
+        for n_jobs in [1, 2, 3, 8, 16]
+    ]
+
+    preds = [e.predict(boston.data[500:, :]) for e in ests]
+    for pred1, pred2 in zip(preds, preds[1:]):
+        assert_array_almost_equal(pred1, pred2)
+    lengths = np.array([[gp.length_ for gp in e._programs[-1]] for e in ests])
+    for len1, len2 in zip(lengths, lengths[1:]):
+        assert_array_almost_equal(len1, len2)
+
+    # Check the transformer
+    ests = [
+        SymbolicTransformer(population_size=100, hall_of_fame=50,
+                            generations=4, n_jobs=n_jobs,
+                            random_state=0).fit(boston.data[:100, :],
+                                                boston.target[:100])
+        for n_jobs in [1, 2, 3, 8, 16]
+    ]
+
+    preds = [e.transform(boston.data[500:, :]) for e in ests]
+    for pred1, pred2 in zip(preds, preds[1:]):
+        assert_array_almost_equal(pred1, pred2)
+    lengths = np.array([[gp.length_ for gp in e._programs[-1]] for e in ests])
+    for len1, len2 in zip(lengths, lengths[1:]):
+        assert_array_almost_equal(len1, len2)
+
+    # Check the classifier
+    ests = [
+        SymbolicClassifier(population_size=100, generations=4, n_jobs=n_jobs,
+                           random_state=0).fit(cancer.data[:100, :],
+                                               cancer.target[:100])
+        for n_jobs in [1, 2, 3, 8, 16]
+    ]
+
+    preds = [e.predict(cancer.data[500:, :]) for e in ests]
+    for pred1, pred2 in zip(preds, preds[1:]):
+        assert_array_almost_equal(pred1, pred2)
+    lengths = np.array([[gp.length_ for gp in e._programs[-1]] for e in ests])
+    for len1, len2 in zip(lengths, lengths[1:]):
+        assert_array_almost_equal(len1, len2)
+
+
 def test_pickle():
     """Check pickability"""
